requirements.txt
matplotlib==3.6.2
numpy==1.24.4
scikit_learn
setuptools==65.5.1
sympy==1.11.1
torch==2.2.2
tqdm==4.66.2
colorama
wget
chess
Pyinstaller
pyyaml
pandas
pykan
